Add unit tests for GastosHandler and GastosRouter; update application properties and refactor Gasto model

+++ src/main/java/com/interview/gastos_api/model/Gasto.java
@@ -10,5 +10,5 @@ public class Gasto {
     private Long id;
     private Long empleadoId;
     private LocalDate fecha;
-    private Long monto;
+    private double monto;
 }
